+++ CHANGELOG.md
@@ -2,7 +2,15 @@
 
 ## [Unreleased](https://github.com/TykTechnologies/tyk-operator/tree/HEAD)
 [Full Changelog](https://github.com/TykTechnologies/tyk-operator/compare/v0.12.0...HEAD)
-- Operator was failing to remove finalizers from ApiDefinition that was already deleted in Dashboard
+
+**Added**
+- Added Basic Authentication support [Issue #534](https://github.com/TykTechnologies/tyk-operator/issues/534)
+
+**Fixed**
+- Operator was failing to remove finalizers from ApiDefinition that was already deleted in Dashboard [#469](https://github.com/TykTechnologies/tyk-operator/issues/469)
+- Fixed the problem of linking existing security policies while migration [#204](https://github.com/TykTechnologies/tyk-operator/issues/204)
+- Fix Security Policy tests
+
 
 **Added**
 - snapshot tool can now export only SecurityPolicy objects without specifying